Only log controller parameters when configured to.
This adds a new config option `log_controller_parameters`, that defaults
to false.  This needs to be set to true to enable logging of the params
hash, otherwise this is omitted.

Rationale:

The change in shadabahmed/logstasher#14 has implications that are not
immediately obvious.  It will cause the elasticsearch mapping for the
parameters field to become a new dynamic mapping that gets indexed.
This can lead to collisions if different apps use the same params key
with different types, which results in log messages being lost.  It also
causes the indexes to become significantly larger (in our case, this
change tripled the size).

In light of this, I think the not logging these fields by default is the
correct choice.