Add option to serialize parameters
By default, LogStasher serializes parameters as a JSON string. This is
done to ensure that elasticsearch can index params from different
applications without risk of collisions from params that are named the
same but typed differently.

This adds an option to disable this behavior by setting
`::LogStasher.serialize_parameters = false`. When set to `false` the
parameters will be sent unserialized so that they can be indexed
individually by elasticsearch.