Add a new #log_as_json method to make it easy to log json messages
If you set LogStasher.metadata, it will be appended to the message with
a :metadata field before calling #to_json.

You can also supply a :as_logstash_event if you wish for your payload to
be wrapped in a LogStash::Event class that adds version and timestamp
fields ready to be consumed by logstash/ elasticsearch.

With this, we were able to clean up the log_subscriber code.